refactor: move use cases to `application` layer
