+++ src/test/java/io/pakland/mdas/githubstats/domain/service/GetOrganizationFromTeamNameTest.java
@@ -1,5 +1,6 @@
 package io.pakland.mdas.githubstats.domain.service;
 
+import io.pakland.mdas.githubstats.application.GetOrganizationFromTeamName;
 import io.pakland.mdas.githubstats.domain.model.Organization;
 import io.pakland.mdas.githubstats.domain.model.Team;
 import io.pakland.mdas.githubstats.domain.repository.TeamRepository;
