+++ src/test/java/io/pakland/mdas/githubstats/domain/service/GetTeamsFromOrganizationTest.java
@@ -1,5 +1,6 @@
 package io.pakland.mdas.githubstats.domain.service;
 
+import io.pakland.mdas.githubstats.application.GetTeamsFromOrganization;
 import io.pakland.mdas.githubstats.domain.model.Organization;
 import io.pakland.mdas.githubstats.domain.model.Team;
 import org.junit.jupiter.api.Test;
